@@ -1,5 +1,5 @@
 {
-  "name": "@jnmorse/josephmorse.me",
+  "name": "josephmorse.me",
   "version": "0.2.0",
   "description": "Personal Website",
   "author": "Joseph Morse <tamed.lionheart@gmail.com> (https://github.com/jnmorse)",
@@ -22,7 +22,7 @@
   "private": true,
   "scripts": {
     "start": "node server.js",
-    "build": "NODE_ENV=production webpack -p --config webpack.production.config.js"
+    "build": "webpack"
   },
   "repository": {
     "type": "git",
@@ -33,42 +33,38 @@
     "email": "tamed.lionheart@gmail.com"
   },
   "devDependencies": {
-    "babel-core": "^6.4.5",
-    "babel-eslint": "^4.1.6",
-    "babel-loader": "^6.2.1",
-    "babel-plugin-transform-runtime": "^6.4.3",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-react-hmre": "^1.0.1",
-    "css-loader": "^0.23.1",
-    "eslint": "^1.10.3",
-    "eslint-plugin-babel": "^3.0.0",
-    "eslint-plugin-react": "^3.16.1",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.8.5",
-    "html-webpack-plugin": "^2.7.1",
-    "node-sass": "^3.4.2",
-    "postcss-loader": "^0.8.0",
-    "react-hot-loader": "^1.3.0",
-    "react-tap-event-plugin": "^0.2.1",
-    "sass-loader": "^3.1.2",
-    "source-map-loader": "^0.1.5",
-    "style-loader": "^0.13.0",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.12.11",
-    "webpack-dev-middleware": "^1.5.1",
-    "webpack-hot-middleware": "^2.6.4"
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-react-hmre": "^1.1.1",
+    "css-loader": "^0.28.5",
+    "eslint": "^4.5.0",
+    "eslint-plugin-react": "^7.2.1",
+    "extract-text-webpack-plugin": "^3.0.0",
+    "file-loader": "^0.11.2",
+    "html-webpack-plugin": "^2.30.1",
+    "node-sass": "^4.5.3",
+    "react-hot-loader": "^1.3.1",
+    "react-tap-event-plugin": "^2.0.1",
+    "webpack": "^3.5.5",
+    "webpack-dev-middleware": "^1.12.0",
+    "webpack-hot-middleware": "^2.18.2"
   },
   "dependencies": {
-    "bootstrap-sass": "^3.3.6",
+    "axios": "^0.16.2",
+    "body-parser": "^1.17.2",
+    "bootstrap-sass": "^3.3.7",
+    "cors": "^2.8.4",
     "express": "^4.15.4",
-    "font-awesome": "^4.5.0",
-    "jquery": "^2.2.0",
-    "node-sass-middleware": "^0.9.7",
-    "radium": "^0.16.5",
-    "react": "^0.14.6",
-    "react-dom": "^0.14.6",
-    "react-router": "^2.0.0-rc5",
-    "sendgrid": "^5.2.1"
+    "font-awesome": "^4.7.0",
+    "jquery": "^3.2.1",
+    "node-sass-middleware": "^0.11.0",
+    "radium": "^0.19.4",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1",
+    "react-router-dom": "^4.1.2",
+    "sendgrid": "^5.2.2"
   }
 }
