@@ -1,6 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
-import { Router, Route, IndexRoute, browserHistory } from 'react-router';
+// import { Router, Route, IndexRoute } from 'react-router';
+import { BrowserRouter, Route, IndexRoute } from 'react-router-dom';
 import Base from './base';
 import Home from './routes/Home';
 import About from './routes/About';
@@ -8,26 +9,28 @@ import Contact from './routes/Contact';
 import Portfolio from './routes/Portfolio/Portfolio';
 
 var WebFontConfig = {
-  google: { families: [ 'Roboto:400,300,500:latin' ] }
+  google: { families: ['Roboto:400,300,500:latin'] }
 };
 
 (function() {
   var wf = document.createElement('script');
-  wf.src = ('https:' == document.location.protocol ? 'https' : 'http') +
-      '://ajax.googleapis.com/ajax/libs/webfont/1/webfont.js';
+  wf.src =
+    ('https:' == document.location.protocol ? 'https' : 'http') +
+    '://ajax.googleapis.com/ajax/libs/webfont/1/webfont.js';
   wf.type = 'text/javascript';
   wf.async = 'true';
   var s = document.getElementsByTagName('script')[0];
   s.parentNode.insertBefore(wf, s);
 })();
 
-render((
-  <Router history={browserHistory}>
-    <Route path="/" component={Base}>
-      <IndexRoute component={Home} />
-      <Route path="about" component={About} />
-      <Route path="portfolio" component={Portfolio} />
-      <Route path="contact" component={Contact} />
-    </Route>
-  </Router>
-), document.querySelector('#app'));
+render(
+  <BrowserRouter>
+    <Base>
+      <Route exact path="/" component={Home} />
+      <Route path="/home" component={Home} />
+      <Route path="/about" component={About} />
+      <Route path="/contact" component={Contact} />
+    </Base>
+  </BrowserRouter>,
+  document.querySelector('#app')
+);
