@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Link } from 'react-router';
+import { Link } from 'react-router-dom';
 import Header from '../elements/Header.jsx';
 
 export default class Home extends Component {
@@ -30,27 +30,46 @@ export default class Home extends Component {
       },
       header: {
         color: '#fafafa',
-        textShadow: '5px 5px 5px #3a3a3a,'
-         + '3px 3px 3px #3a3a3a,'
-         + '1px 1px 1px #3a3a3a'
+        textShadow:
+          '5px 5px 5px #3a3a3a,' +
+          '3px 3px 3px #3a3a3a,' +
+          '1px 1px 1px #3a3a3a'
       }
     };
   }
 
-  render () {
+  render() {
     return (
       <div>
-        <section style={Object.assign({}, this.state.style, this.state.section1)} id="home" className="container-fluid">
-          <div style={{position: 'absolute', top: '50%', left: '50%', transform: 'translate(-50%, -50%)'}}>
+        <section
+          style={Object.assign({}, this.state.style, this.state.section1)}
+          id="home"
+          className="container-fluid"
+        >
+          <div
+            style={{
+              position: 'absolute',
+              top: '50%',
+              left: '50%',
+              transform: 'translate(-50%, -50%)',
+              textAlign: 'center'
+            }}
+          >
             <Header title="Joseph Morse" style={this.state.header}>
               <h2>Better Code To Cover Your Backend</h2>
             </Header>
             <div>
-              <Link to="/contact" style={this.state.button} className="btn btn-primary btn-lg">Contact Me</Link>
-              <Link to="/portfolio" style={this.state.button} className="btn btn-default btn-lg">See My Work</Link>
+              <Link
+                to="/contact"
+                style={this.state.button}
+                className="btn btn-primary btn-lg"
+              >
+                Contact Me
+              </Link>
             </div>
           </div>
-        </section>{/* Hero */}
+        </section>
+        {/* Hero */}
       </div>
     );
   }
