@@ -1,31 +1,30 @@
-var webpack = require('webpack');
-var path = require('path');
-
-var app = path.join(__dirname, 'app');
+var webpack = require("webpack");
 
 module.exports = {
   entry: {
-    app: path.join(__dirname, 'app', 'scripts', 'main')
+    app: `${__dirname}/app/scripts/main`
   },
   output: {
-    path: path.join(__dirname, 'dist'),
-    filename: '[name].js',
-    // publicPath: '/'
+    path: __dirname,
+    filename: "[name].js",
+    publicPath: "/"
   },
   module: {
     loaders: [
       {
         test: /\.jsx?$/,
         exclude: /node_modules/,
-        loaders: ['babel']
-
+        loaders: ["babel-loader"]
       }
     ]
   },
   resolve: {
-    extensions: ['', '.js', '.jsx']
+    extensions: [".js", ".jsx"]
   },
   plugins: [
-    new webpack.optimize.CommonsChunkPlugin(['app', 'vendor'], 'js/[name].js'),
+    new webpack.optimize.CommonsChunkPlugin({
+      name: "vendor",
+      filename: "vendor.js"
+    })
   ]
 };
